requirements.txt
# Core PyTorch and audio processing
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
librosa>=0.10.0
soundfile>=0.12.1

# Data handling and datasets
datasets>=2.14.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Configuration and CLI
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.65.0

# Evaluation metrics
jiwer>=3.0.0
scikit-learn>=1.3.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# Distributed training (optional)
accelerate>=0.24.0
deepspeed>=0.10.0
fairscale>=0.4.0

# Speech processing utilities
python-speech-features>=0.6
pesq>=0.0.4

# Optimization
torch-optimizer>=0.3.0

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Data processing
requests>=2.28.0
urllib3>=1.26.0